Lots of UI changes. One of which required some rearranging of code. - almost universal toolbar (on all but the main page). Instead of the floating “go back”, it’s the toolbar. - Removed the fetch button from the map page and a little button on   the front page. - added the change notifications for the properties in SFOOrganization   to clear the warnings. - Removed the TranslatorHeader, since it’s only used in that one page - Layout fixes in the Festival schedule. Still more to go.
TextField handling:
We were using 2 methods to get the text data (event filter and
validator). The event filter doesn’t work with TextField, which we’re
using everywhere now. The validator takes the model object that its
validating for, but since that’s passed in through QML, the model needs
to be a QVariant. They share a base class, but that’s only used in C++.
So, in addition to SFOScheduleModel, SFOTranslateController also
implements SFOValidatorReceiver.