Another pass at the Add Phrase feature. Less options, aiming to be less confusing.
Now the GUI only accepts Japanese to English translations. Submit
button is only enabled on proper input.

Provided a network accessible API down to FJClient, exposed through
SFOContext. Cleaned up some of the SFOSubmitWordModel code. (I think
this only sees if there is a network interface, not even if it’s on.
But it’s a rare case, so we might let this go for now.)

Moved finding the input language to SFOTypes, so other classes can use
it.

Added SFOSubmitWordModel to the lupdate sources.