Shuffling of the other two apps. Partners Map is now Taste of Osaka, and partners that aren’t restaurants are moved into a new Partners List. This replaces the Sister City app, which just showed their website.
Taste of Osaka: This is the old map, but with the names removed, and
only the restaurants. Instead of going to the detail page on tap, we
show the name in a little popup. Tapping the popup will show the
detail page.

Partners list is similar to the schedule, but with non-food partners.
Partners with a URL will have a link off the list to their website.

Under the hood:
SFOItemModel was the first bit written, and its name was too general.
We made it more general this time, and displays any partner, and is
used for the map (Taste of Osaka) and Partners. Since we’re using
Category more heavily (still hardcoded), that has been put into
SFOTypes. It handles all the visible elements that we currently want to
display.

SFOController is the new global controller. It handles updates from
the SFOContext and sends it to the SFOItemModel instances. (We leave
the other controllers the way they were. Moved the refresh into here,
now that we have a central place.

Small tweak: SFOTypes is now a class, and it holds some global instance
variables and functions (well, just one).