WrapAnywhere was used because Japanese text wasn’t wrapping correctly earlier, but it was working with setting width, so go back to Wrap.
Added some padding in the large text blocks. TextArea’s were white, 
which separated the text, so this helps separated it visually.

Tweaked the width and text clipping in the partner list to give the
image/logo more room.

Removed resetting the model (emitting dataChanged) because that was
causing a crash (and it was redundant).