Some tweaks to streamline building. Mainly in the SFOsaka.pro
Build 21, Ver 1.6

FJClient: Separate the errors with empty and invalid JSON
SFOContext: Display the full host in the GUI in debug. Use #QT_DEBUG
to switch between dev and live servers (Need to add beta at some point,
but we may not have that this time…)

SFOsaka.pro:
Finally figured out the CONFIG() usage, so no more tweaking the pro file
for different builds. The library can be determined for debug/release
for all platforms.