Added the SOYNET Highlight section to the app. It’s a subclass of the Organization class, like Partners and Restaurants, and will just list some items.
We can play with the interface, like making the first item more
prominent. But, this is just to get the basic data pulled from the
server and cached like everything else.

SFOItemModel was holding the Partner. Now it holds the more abstract
Organization. If more specific information is requested from the model,
it will downcast to the specific type.

Most changes were in SFOContext.cpp, which doesn’t do as much template
magic, since we don’t care what type of Organization we’re holding
anymore. And, like the other Organization subclasses, we cache the
AppHighlight in its own JSON file locally.

Cleanup: Index from Qt::UserRole in the roles, rather than from 0. This
wasn’t causing problems that I know of, but maybe there were problems
debugging earlier. This is is just the better way to do it.
maybe