Feature updates for 2018 and fixes for iOS App Store submission.
This is a monster change, but, since we had some feature requests
come in later than we’d like, a lot of things got crammed into this
one change.

The app store submissions now require the asset catalog, so all the
icons were moved into Images.xcassets and Contents.json and out of the
the qmake project file and Info.plist.

QML is a lot more picky about using an Item’s anchors property vs the
Layout property, even though, it seems like some things still rely on
the anchor, which result in warnings. I’m not sure how to get around
that yet.

Feature List:
SOYNET won’t have a booth this year, so changed that item to the
scavenger hunt. Added a ‘transportation’ item. Since these are identical
items, changed HighlightItem.qml to GenericItem.qml. They use a
different list model, so, to avoid changing too much qml, we just use
a different list model, where they only differ with the ‘model’
property.

The most work on the client side was in the main page, since a fifth
item, on the iPhone 5 size screen makes the screen too big. So, changed
that little area to scroll. This was tricky, so I added a lot more 
comments.

Server side: Added a Transportation model, and the appropriate view and
url support for it. Changed the modificationDate to be auto incremented.
This was annoying to manually do it every time we changed a field.